Compile for Swift 6.1
